backend/requirements.txt
# ── Web framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1

# ── Database ──────────────────────────────────────────────
sqlalchemy==2.0.30
pymysql==1.1.1
alembic==1.13.1          # optional — for migrations

# ── Validation ────────────────────────────────────────────
pydantic[email]==2.7.1

# ── Auth ──────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9  # required for OAuth2PasswordBearer form parsing

# ── Env / Config ──────────────────────────────────────────
python-dotenv==1.0.1     # optional — load .env files
